backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.7
dj-database-url
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
qrcode==7.4.2
requests==2.31.0
